+++ apps/mcp-endpoint-shield/github-cli-hooks/akto-hooks.py
@@ -6,7 +6,8 @@ import sys
 if not os.getenv("LOG_DIR"):
     os.environ["LOG_DIR"] = os.path.expanduser("~/akto/.github/akto/vscode/logs")
 
-from akto_ingestion_utility import run_agent_stop_hook, run_observability_hook
+from akto_ingestion_utility import run_observability_hook
+from akto_copilot_common import run_agent_stop_hook
 
 if __name__ == "__main__":
     if len(sys.argv) < 2:
