+++ apps/mcp-endpoint-shield/github-cli-hooks/hooks.json
@@ -28,48 +28,84 @@
         "timeoutSec": 30
       }
     ],
-    "_disabled_SessionStart": [
+    "sessionStart": [
       {
         "type": "command",
-        "bash": "bash ./.github/hooks/akto-hook-wrapper.sh akto-hooks.py SessionStart",
-        "powershell": "powershell -ExecutionPolicy Bypass -File .github/hooks/akto-hook-wrapper.ps1 akto-hooks.py SessionStart",
-        "comment": "DISABLED: not supported by current GitHub Copilot hook contract. To re-enable, rename key back to 'SessionStart' and ship akto-hook-wrapper.{sh,ps1} + akto-hooks.py.",
+        "bash": "bash ./.github/hooks/akto-hook-wrapper.sh akto-hooks.py sessionStart",
+        "powershell": "powershell -ExecutionPolicy Bypass -File .github/hooks/akto-hook-wrapper.ps1 akto-hooks.py sessionStart",
+        "comment": "Ingest session start event to Akto for observability",
         "timeoutSec": 10
       }
     ],
-    "_disabled_Stop": [
+    "sessionEnd": [
       {
         "type": "command",
-        "bash": "bash ./.github/hooks/akto-hook-wrapper.sh akto-hooks.py Stop",
-        "powershell": "powershell -ExecutionPolicy Bypass -File .github/hooks/akto-hook-wrapper.ps1 akto-hooks.py Stop",
-        "comment": "DISABLED: not supported by current GitHub Copilot hook contract. To re-enable, rename key back to 'Stop' and ship akto-hook-wrapper.{sh,ps1} + akto-hooks.py.",
+        "bash": "bash ./.github/hooks/akto-hook-wrapper.sh akto-hooks.py sessionEnd",
+        "powershell": "powershell -ExecutionPolicy Bypass -File .github/hooks/akto-hook-wrapper.ps1 akto-hooks.py sessionEnd",
+        "comment": "Ingest session end event to Akto for observability",
         "timeoutSec": 10
       }
     ],
-    "_disabled_PreCompact": [
+    "agentStop": [
       {
         "type": "command",
-        "bash": "bash ./.github/hooks/akto-hook-wrapper.sh akto-hooks.py PreCompact",
-        "powershell": "powershell -ExecutionPolicy Bypass -File .github/hooks/akto-hook-wrapper.ps1 akto-hooks.py PreCompact",
-        "comment": "DISABLED: not supported by current GitHub Copilot hook contract. To re-enable, rename key back to 'PreCompact' and ship akto-hook-wrapper.{sh,ps1} + akto-hooks.py.",
+        "bash": "bash ./.github/hooks/akto-hook-wrapper.sh akto-hooks.py agentStop",
+        "powershell": "powershell -ExecutionPolicy Bypass -File .github/hooks/akto-hook-wrapper.ps1 akto-hooks.py agentStop",
+        "comment": "Ingest turn-completion event to Akto for observability",
         "timeoutSec": 10
       }
     ],
-    "_disabled_SubagentStart": [
+    "subagentStart": [
       {
         "type": "command",
-        "bash": "bash ./.github/hooks/akto-hook-wrapper.sh akto-hooks.py SubagentStart",
-        "powershell": "powershell -ExecutionPolicy Bypass -File .github/hooks/akto-hook-wrapper.ps1 akto-hooks.py SubagentStart",
-        "comment": "DISABLED: not supported by current GitHub Copilot hook contract. To re-enable, rename key back to 'SubagentStart' and ship akto-hook-wrapper.{sh,ps1} + akto-hooks.py.",
+        "bash": "bash ./.github/hooks/akto-hook-wrapper.sh akto-hooks.py subagentStart",
+        "powershell": "powershell -ExecutionPolicy Bypass -File .github/hooks/akto-hook-wrapper.ps1 akto-hooks.py subagentStart",
+        "comment": "Ingest subagent-start event to Akto for observability",
         "timeoutSec": 10
       }
     ],
-    "_disabled_SubagentStop": [
+    "subagentStop": [
       {
         "type": "command",
-        "bash": "bash ./.github/hooks/akto-hook-wrapper.sh akto-hooks.py SubagentStop",
-        "powershell": "powershell -ExecutionPolicy Bypass -File .github/hooks/akto-hook-wrapper.ps1 akto-hooks.py SubagentStop",
-        "comment": "DISABLED: not supported by current GitHub Copilot hook contract. To re-enable, rename key back to 'SubagentStop' and ship akto-hook-wrapper.{sh,ps1} + akto-hooks.py.",
+        "bash": "bash ./.github/hooks/akto-hook-wrapper.sh akto-hooks.py subagentStop",
+        "powershell": "powershell -ExecutionPolicy Bypass -File .github/hooks/akto-hook-wrapper.ps1 akto-hooks.py subagentStop",
+        "comment": "Ingest subagent-completion event to Akto for observability",
+        "timeoutSec": 10
+      }
+    ],
+    "errorOccurred": [
+      {
+        "type": "command",
+        "bash": "bash ./.github/hooks/akto-hook-wrapper.sh akto-hooks.py errorOccurred",
+        "powershell": "powershell -ExecutionPolicy Bypass -File .github/hooks/akto-hook-wrapper.ps1 akto-hooks.py errorOccurred",
+        "comment": "Ingest agent error event to Akto for observability",
+        "timeoutSec": 10
+      }
+    ],
+    "preCompact": [
+      {
+        "type": "command",
+        "bash": "bash ./.github/hooks/akto-hook-wrapper.sh akto-hooks.py preCompact",
+        "powershell": "powershell -ExecutionPolicy Bypass -File .github/hooks/akto-hook-wrapper.ps1 akto-hooks.py preCompact",
+        "comment": "Ingest pre-compact event to Akto for observability",
+        "timeoutSec": 10
+      }
+    ],
+    "notification": [
+      {
+        "type": "command",
+        "bash": "bash ./.github/hooks/akto-hook-wrapper.sh akto-hooks.py notification",
+        "powershell": "powershell -ExecutionPolicy Bypass -File .github/hooks/akto-hook-wrapper.ps1 akto-hooks.py notification",
+        "comment": "Ingest CLI notification event to Akto for observability",
+        "timeoutSec": 10
+      }
+    ],
+    "userPromptTransformed": [
+      {
+        "type": "command",
+        "bash": "bash ./.github/hooks/akto-hook-wrapper.sh akto-hooks.py userPromptTransformed",
+        "powershell": "powershell -ExecutionPolicy Bypass -File .github/hooks/akto-hook-wrapper.ps1 akto-hooks.py userPromptTransformed",
+        "comment": "Ingest transformed-prompt event to Akto for observability",
         "timeoutSec": 10
       }
     ]
