+++ apps/mcp-endpoint-shield/shared/akto_ingestion_utility.py
@@ -9,7 +9,7 @@ import ssl
 import sys
 import time
 import urllib.request
-from typing import Any, Dict, List, Optional, Tuple, Union
+from typing import Any, Dict, List, Optional, Union
 
 try:
     from akto_machine_id import get_machine_id, get_username
@@ -259,12 +259,12 @@ def _id_fields(fm: Dict[str, Any]) -> List[str]:
 
 
 # Terminal `copilot` CLI sends camelCase; VS Code's Copilot Chat sends snake_case.
-_CAMEL_TO_SNAKE_ALIASES = {"sessionId": "session_id", "transcriptPath": "transcript_path"}
+INPUT_FIELD_TRANSFORMATIONS = {"sessionId": "session_id", "transcriptPath": "transcript_path"}
 
 
 def _alias_camel_keys(input_data: Dict[str, Any]) -> Dict[str, Any]:
     """Fill in the snake_case key from its camelCase alias when only the latter is present."""
-    for camel, snake in _CAMEL_TO_SNAKE_ALIASES.items():
+    for camel, snake in INPUT_FIELD_TRANSFORMATIONS.items():
         if snake not in input_data and camel in input_data:
             input_data[snake] = input_data[camel]
     return input_data
@@ -554,87 +554,6 @@ def send_ingestion_data(
     except Exception as e:
         logger.error(f"Ingestion error: {e}")
         return None
-
-
-# ── Copilot CLI transcript (agentStop) ──────────────────────────────────────────
-# agentStop has no inline response text — only a transcript_path to events.jsonl.
-
-def _extract_last_turn(transcript_path: str) -> Tuple[str, str]:
-    """Read events.jsonl and return (last user prompt, that turn's assistant response)."""
-    if not transcript_path or not os.path.exists(transcript_path):
-        return "", ""
-    last_user = ""
-    response_parts: List[str] = []
-    try:
-        with open(transcript_path, encoding="utf-8") as f:
-            for line in f:
-                line = line.strip()
-                if not line:
-                    continue
-                try:
-                    entry = json.loads(line)
-                except json.JSONDecodeError:
-                    continue
-                data = entry.get("data", {}) or {}
-                if entry.get("type") == "user.message":
-                    content = data.get("content", "")
-                    if content:
-                        last_user = content
-                        response_parts = []  # new turn started — drop the prior one
-                elif entry.get("type") == "assistant.message":
-                    content = data.get("content", "")
-                    if content:
-                        response_parts.append(content)
-    except OSError:
-        pass
-    return last_user, "".join(response_parts)
-
-
-def run_agent_stop_hook() -> None:
-    """Read the turn's transcript and ingest it; falls back to metadata-only if empty."""
-    logger = setup_logger("hook-executions.log")
-    logger.info("=== agentStop hook started ===")
-    try:
-        input_data = _alias_camel_keys(json.load(sys.stdin))
-        logger.info("agentStop input:\n%s", json.dumps(input_data, indent=2))
-        session_info = resolve_session_info(input_data, logger)
-        transcript_path = os.path.expanduser(input_data.get("transcript_path", ""))
-        user_prompt, response_text = _extract_last_turn(transcript_path)
-
-        # transcript write can lag the agentStop event by a beat — retry briefly.
-        for _ in range(5):
-            if response_text or not user_prompt:
-                break
-            time.sleep(0.2)
-            user_prompt, response_text = _extract_last_turn(transcript_path)
-
-        if user_prompt or response_text:
-            logger.info(f"Extracted turn — prompt: {len(user_prompt)} chars, response: {len(response_text)} chars")
-            send_ingestion_data(
-                hook_name="agentStop",
-                request_payload=user_prompt,
-                response_payload=response_text,
-                session_info=session_info,
-                input_data=input_data,
-                guardrails=False,
-                logger=logger,
-            )
-        else:
-            logger.info("No conversational content found in transcript — ingesting metadata only")
-            send_ingestion_data(
-                hook_name="agentStop",
-                request_payload=input_data,
-                response_payload={},
-                session_info=session_info,
-                input_data=input_data,
-                guardrails=False,
-                logger=logger,
-            )
-        logger.info("=== agentStop hook completed ===")
-    except Exception as e:
-        logger.error(f"Main error: {e}")
-    print(json.dumps({}))
-    sys.exit(0)
 
 
 # ── Hook runners ──────────────────────────────────────────────────────────────
